googhud 1.1
-new customization menu
-new customization folder with many options such as health, chat, scoreboard, etc.
-new and better file layouts
-many color updates to various parts
-better resolution compatibility
-many many many many fixes

this update is the result of a ton of work from myself individually, i am still working on this more or less alone. thanks to the folks that helped me with testing though!!!

this is still not even close to where i want this hud to be, and more updates are to come. major plans include:
-more and better customization folder options
-more and better customization menu options
-tf2hudeditor integration
-medigun type icons
-probably some more stuff im forgetting because im pushing this update at 4:00 AM